requirements.txt
fastapi
uvicorn
httpx
pydantic
pydantic-settings
python-dotenv
openai-whisper
numpy
websockets